requirement.txt
streamlit_lottie==0.0.3
requests==2.28.1
streamlit==1.15.1
Pillow==9.3.0